requirements.txt
mysql-connector-python==8.2.0
numpy==1.26.3
pandas==2.1.4
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tk==0.1.0
tzdata==2023.4
